Implement CustomizationMenuBase for sticker customization
- Created a new CustomizationMenuBase component to handle shared customization logic.
- Refactored StickerCustomizationMenu to utilize CustomizationMenuBase, simplifying the component structure.
- Removed redundant state management and event handling from StickerCustomizationMenu.
- Added responsive navigation and subcategory tabs to the new base component.
- Integrated export and download functionality into CustomizationMenuBase.
